♻️ Refactor: Form 로직 변경
