+++ src/shared/ui/form/Form.tsx
@@ -1,36 +1,8 @@
-import type {
-  SubmitHandler,
-  UseFormHandleSubmit,
-  FieldValues,
-} from 'react-hook-form'
-interface FormProps {
-  handleSubmit: UseFormHandleSubmit<FieldValues>
-  children: React.ReactNode
-  setError?: any
-  apiRequest?: any
-  className?: string
-}
+import type { FormProps } from './Form.type'
 
-export default function Form({
-  apiRequest,
-  handleSubmit,
-  setError,
-  children,
-  ...props
-}: FormProps) {
-  const onSubmit: SubmitHandler<FieldValues> = async (data) => {
-    try {
-      await apiRequest(data, setError)
-    } catch (error) {
-      console.error(error)
-    }
-  }
-
+export default function Form({ onSubmit, children, ...props }: FormProps) {
   return (
-    <form
-      onSubmit={handleSubmit(onSubmit)}
-      className={`${props.className || ''}`}
-    >
+    <form onSubmit={onSubmit} className={`${props.className || ''}`}>
       {children}
     </form>
   )
