+++ src/shared/ui/form/Form.type.ts
@@ -1,0 +1,5 @@
+export interface FormProps {
+  onSubmit: (event: React.FormEvent<HTMLFormElement>) => void
+  children: React.ReactNode
+  className?: string
+}
